% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{labs.colors}
\alias{labs.colors}
\title{Color ramp based on \code{mblabs} theme colors}
\usage{
labs.colors(x, colors = unname(pal), ...)
}
\arguments{
\item{x}{number of colors to interpolate}

\item{colors}{colors to interpolate; must be a valid argument to
    \code{\link[grDevices]{col2rgb}()}.}

\item{...}{
  Arguments passed on to \code{\link[grDevices:colorRamp]{grDevices::colorRamp}}
  \describe{
    \item{\code{bias}}{a positive number.  Higher values give more widely spaced
    colors at the high end.}
    \item{\code{space}}{a character string; interpolation in RGB or CIE Lab
    color spaces.}
    \item{\code{interpolate}}{use spline or linear interpolation.}
    \item{\code{alpha}}{logical: should alpha channel (opacity) values be
    returned?   It is an error to give a true value if
    \code{space} is specified.}
  }}
}
\value{
A function to interpolate colors
}
\description{
Default qualitative color ramp derived from \code{mblabs} Bootstrap theme.
}
\examples{
x <- rchisq(100, df=4)
hist(x, freq=FALSE, ylim=c(0, 0.2), col=labs.colors(20), border="white")
hist(x, freq=FALSE, ylim=c(0, 0.2), col=labs.colors(8), border="white")

}
