% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{pal}
\alias{pal}
\title{MBLABS 12-color palette}
\format{
A named character vector of hex color codes
}
\usage{
pal
}
\description{
Custom named qualitative color palette with 12 hues and 1 black derived from
\code{mblabs} Bootstrap theme.
}
\examples{
pie(rep(1, length(pal)), col=pal, labels=names(pal), border="white", clockwise=TRUE)

}
\keyword{datasets}
